Add test for EventSummary
